+++ remotestate-ts/src/lib/index.ts
@@ -14,7 +14,7 @@ export type {
   LocalQueryHandlers,
   LocalStateClientOptions,
 } from "./local";
-export type { Path, PathLike, PathSegment, RelativePath } from "./path";
+export type { Path, PathLike, PathSegment } from "./path";
 export type {
   IncomingMessage,
   OutgoingMessage,